% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get-helpers.R
\name{check_db}
\alias{check_db}
\title{Check user-specified database name and transforms it (if needed).
If the user improperly specified a database,
a message tells them how to fix it.}
\usage{
check_db(db, verbose = FALSE)
}
\arguments{
\item{db}{A string specifying the database to query. Defaults to
\code{NULL}, in which case the most recent database is queried. These
names should be specified either as a four digit year (e.g., 1718) or as
in the database, e.g., \code{"ORExt1718"}
would query the 1718 database.}

\item{verbose}{Do you want the format explained when acceptable, but non-exact matching strings, are used?}
}
\value{
Character vector — correctly formatted database name.
}
\description{
Check user-specified database name and transforms it (if needed).
If the user improperly specified a database,
a message tells them how to fix it.
}
\examples{
library(orextdb)
\dontrun{
check_db(db = current_db())
}
}
