% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get-helpers.R
\name{check_tables}
\alias{check_tables}
\title{Function to check table name}
\usage{
check_tables(tbl)
}
\value{
Warning or nothing
}
\description{
Stops R, returns an error, and lists available tables, if an invalid
table name is specified.
}
\keyword{internal}
