% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get-helpers.R
\name{create_empty_frame}
\alias{create_empty_frame}
\title{Create empty data frame with correct column names.}
\usage{
create_empty_frame(table, db)
}
\arguments{
\item{table}{A string specifying the specific table from the Oregon Extended
live database. Should be one of "Accommodations", "Answers", "Districts",
"Exams", "Items", "Preferences", "Schools", "Students", "Students_old",
"Submissions", "SupplementalDistricts", "SupplementalSchools", "Tasks",
"User", "UserStudents", or "UserStudents_old".}

\item{db}{A string specifying the database to query. Defaults to
\code{NULL}, in which case the most recent database is queried. These
names should be specified either as a four digit year (e.g., 1718) or as
in the database, e.g., \code{"ORExt1718"}
would query the 1718 database.}
}
\value{
data.frame - contains column names that would be present for user-requested table
}
\description{
This function helps handle situations with no
data by returning an empty data frame with the
correct column names for the user-specified table.
}
\keyword{internal}
