% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get-helpers.R
\name{create_item_table}
\alias{create_item_table}
\title{Convert text to data frame for items}
\usage{
create_item_table(txt)
}
\arguments{
\item{txt}{The raw text data returned from \code{\link{get_raw_data}}}
}
\value{
data.frame - 4 columns: item_id, standard, brt_item_id, and item_difficulty.
}
\description{
Convert text to data frame for items
}
\keyword{internal}
