% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get-helpers.R
\name{current_db}
\alias{current_db}
\title{Determines the current db based on the system date
Creates a string with the name of the db}
\usage{
current_db()
}
\value{
Character vector — the most recent database name.
}
\description{
Determines the current db based on the system date
Creates a string with the name of the db
}
\examples{
library(orextdb)
\dontrun{
current_db()
}
}
