% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get.R
\name{db_get}
\alias{db_get}
\title{Function to query the live OR Extended database and return specific tables}
\usage{
db_get(table, db = NULL, raw = FALSE, key = db_key())
}
\arguments{
\item{table}{A string specifying the specific table from the Oregon Extended
live database. Should be one of "Accomodations", "Answers", "Districts",
"Exams", "Items", "Preferences", "Schools", "Students", "Students_old",
"Submissions", "SupplementalDistricts", "SupplementalSchools", "Tasks",
"User", "UserStudents", or "UserStudents_old".}

\item{db}{A string specifying the database to query. Defaults to
\code{NULL}, in which case the most recent database is queried. These
names should be specified either as a four digit year (e.g., 1718) or as
in the database, e.g., \code{"ORExt1718"}
would query the 1718 database.}

\item{raw}{Logical, defaults to \code{FALSE}. Should the original tables
from the database be returned? If \code{FALSE}, cleaned up names are
returned and, for the \code{"Items"} table, item difficulties are
returned instead of the full item attributes.}

\item{key}{Your personal key for accessing the Oregon Extended Database.
Defaults to \code{\link[=db_key]{db_key()}}.}
}
\description{
Function to query the live OR Extended database and return specific tables
}
