% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{db_key}
\alias{db_key}
\title{Access your key from .Renviron}
\usage{
db_key()
}
\value{
Either your api key or a message explaining how to load it
}
\description{
This function is primarily used as the default argument for \code{\link[=db_get]{db_get()}}, but
can be run at any time to access your specific key. Note that you should
first set your key with \code{\link[=db_set_key]{db_set_key()}}.
}
\examples{

\dontrun{
library(orextdb)
db_key()
}

}
