% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{db_set_key}
\alias{db_set_key}
\title{Function to set a key for accessing the live OR Extended database}
\usage{
db_set_key(key, overwrite = FALSE)
}
\arguments{
\item{key}{A string of your specific key}

\item{overwrite}{Defaults to \code{FALSE}. If there is a key already in the
.Renviron, should it be overwritten?}
}
\description{
This function adds your key to your .Renviron so it is accessible by
\code{\link[=db_key]{db_key()}}. You should only need to run this function once. Make sure to
restart your local R session after running this function for the changes
to take effect.
}
\examples{
library(orextdb)

\dontrun{
db_set_key("abcdefghikjlmnop")
}
}
