% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{explain_db_format}
\alias{explain_db_format}
\title{explain how to format the database specification when misspecified}
\usage{
explain_db_format(type, reason)
}
\arguments{
\item{type}{= character vector indicating either c('message', 'stop')
which passes to respective functions}

\item{reason}{= character vector indicating the way they misspecified
(changes Note 1 of 2)}
}
\value{
message about database formatting
}
\description{
explain how to format the database specification when misspecified
}
\keyword{internal}
