% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get-helpers.R
\name{get_raw_data}
\alias{get_raw_data}
\title{Return the raw data from the api}
\usage{
get_raw_data(table, db, key)
}
\arguments{
\item{table}{A string specifying the specific table from the Oregon Extended
live database. Should be one of "Accommodations", "Answers", "Districts",
"Exams", "Items", "Preferences", "Schools", "Students", "Students_old",
"Submissions", "SupplementalDistricts", "SupplementalSchools", "Tasks",
"User", "UserStudents", or "UserStudents_old".}

\item{db}{A string specifying the database to query. Defaults to
\code{NULL}, in which case the most recent database is queried. These
names should be specified either as a four digit year (e.g., 1718) or as
in the database, e.g., \code{"ORExt1718"}
would query the 1718 database.}

\item{key}{Your personal key for accessing the Oregon Extended Database.
Defaults to \code{\link[=db_key]{db_key()}}.}
}
\value{
Character vector — the raw data from the API
}
\description{
The raw data is returned as a single string. It then needs to be passed
to \code{\link{parse_txt_data}} or \code{\link{create_item_table}} (if
item data are being returned).
}
\keyword{internal}
