% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%p\%}
\alias{\%p\%}
\title{internal paste function
Note: created by DA and not currently used
Infix function to paste long strings together a little easier}
\usage{
lhs \%p\% rhs
}
\arguments{
\item{lhs}{= character vector to paste on left}

\item{rhs}{= character vector to paste on right}
}
\value{
character vector comprised of \code{lhs} & \code{rhs}
}
\description{
internal paste function
Note: created by DA and not currently used
Infix function to paste long strings together a little easier
}
\keyword{internal}
