% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orextdb-package.R
\docType{package}
\name{orextdb-package}
\alias{orextdb}
\alias{orextdb-package}
\title{orextdb: Retrieves data from the ORExt live database}
\description{
This package downloads any table from the ORExt live database through a secure web API. The `db_get()` function can be use to download any table after first setting a key through the `db_set_key()` function.
}
\author{
\strong{Maintainer}: Daniel Anderson \email{daniela@uoregon.edu} (\href{https://orcid.org/0000-0003-4699-4680}{ORCID})

}
\keyword{internal}
