% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get-helpers.R
\name{return_camelCase_names}
\alias{return_camelCase_names}
\title{Return the names in camelCase}
\usage{
return_camelCase_names(table)
}
\arguments{
\item{table}{A string specifying the specific table from the Oregon Extended
live database. Should be one of "Accommodations", "Answers", "Districts",
"Exams", "Items", "Preferences", "Schools", "Students", "Students_old",
"Submissions", "SupplementalDistricts", "SupplementalSchools", "Tasks",
"User", "UserStudents", or "UserStudents_old".}
}
\value{
A character vector of the new cleaned up names
}
\description{
Return the names in camelCase
}
\keyword{internal}
