% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get-helpers.R
\name{return_snake_case_names}
\alias{return_snake_case_names}
\title{Returns the column names for a given table in snake_case}
\usage{
return_snake_case_names(table)
}
\arguments{
\item{table}{A string specifying the specific table from the Oregon Extended
live database. Should be one of "Accommodations", "Answers", "Districts",
"Exams", "Items", "Preferences", "Schools", "Students", "Students_old",
"Submissions", "SupplementalDistricts", "SupplementalSchools", "Tasks",
"User", "UserStudents", or "UserStudents_old".}
}
\value{
Character vector - formatted column names in table
}
\description{
Returns the column names for a given table in snake_case
}
\keyword{internal}
