% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get-helpers.R
\name{check_db}
\alias{check_db}
\title{Checks the database argument and transforms it if needed}
\usage{
check_db(db)
}
\arguments{
\item{db}{A string specifying the database to query. Defaults to
\code{NULL}, in which case the most recent database is queried. These
names should be specified either as a four digit year (e.g., 1718) or as
in the database, e.g., \code{"ORExt1718"}
would query the 1718 database.}
}
\description{
Checks the database argument and transforms it if needed
}
