% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_db}
\alias{check_db}
\title{Checks the database argument and transforms it if needed}
\usage{
check_db(db)
}
\arguments{
\item{db}{A string specifying the database to query. Defaults to
\code{NULL}, in which case teh most recent database is queried. These
names should be specified as in the database, e.g., \code{"ORExt1920"}
would query the 1920 database.}
}
\description{
Checks the database argument and transforms it if needed
}
